refactor: metadata refactor
