+++ src/lib/state/dataParser.ts
@@ -3,10 +3,10 @@ import relicMainAffixes from 'data/relic_main_affixes.json' with { type: 'json' 
 import relicSubAffixes from 'data/relic_sub_affixes.json' with { type: 'json' }
 import { Constants, Parts, PartsMainStats, Sets, SetsRelics, Stats } from 'lib/constants/constants'
 import { SortOption } from 'lib/optimization/sortOptions'
-import { ScoringMetadata } from 'lib/scoring/characterScorer'
-import DB, { DBMetadata, DBMetadataCharacter, DBMetadataLightCone, DBMetadataSets } from 'lib/state/db'
+import DB from 'lib/state/db'
 import { PresetEffects } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { MetadataCharacter } from 'types/character'
+import { DBMetadata, DBMetadataCharacter, DBMetadataLightCone, DBMetadataSets, ScoringMetadata } from 'types/metadata'
 
 // FIXME HIGH
