+++ src/lib/state/db.ts
@@ -16,7 +16,6 @@ import { getDefaultForm } from 'lib/optimization/defaultForm'
 import { DefaultSettingOptions, SettingOptions } from 'lib/overlays/drawers/SettingsDrawer'
 import { RelicAugmenter } from 'lib/relics/relicAugmenter'
 import { Themes } from 'lib/rendering/theme'
-import { ScoringMetadata, SimulationMetadata } from 'lib/scoring/characterScorer'
 import { oldCharacterScoringMetadata } from 'lib/scoring/oldCharacterScoringMetadata'
 import { SaveState } from 'lib/state/saveState'
 import { ComboState } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
@@ -27,6 +26,7 @@ import { TsUtils } from 'lib/utils/TsUtils'
 import { Utils } from 'lib/utils/utils'
 import { Character } from 'types/character'
 import { Form } from 'types/form'
+import { DBMetadata, ScoringMetadata, SimulationMetadata } from 'types/metadata'
 import { Relic, Stat } from 'types/relic'
 import { CustomPortrait, HsrOptimizerSaveFormat, HsrOptimizerStore, SavedSession, UserSettings } from 'types/store'
 import { create } from 'zustand'
@@ -258,69 +258,6 @@ window.store = create((set) => {
   }
   return store
 })
-
-export type DBMetadataCharacter = {
-  id: string
-  name: string
-  rarity: number
-  path: string
-  element: string
-  max_sp: number
-  stats: Record<string, number>
-  unreleased: boolean
-  traces: Record<string, number>
-  imageCenter: {
-    x: number
-    y: number
-    z: number
-  }
-  displayName: string
-  scoringMetadata: ScoringMetadata
-}
-
-export type DBMetadataLightCone = {
-  id: string
-  name: string
-  rarity: number
-  path: string
-  stats: Record<string, number>
-  unreleased: boolean
-  superimpositions: Record<number, Record<string, number>>
-  displayName: string
-  imageCenter: number
-}
-
-export type DBMetadataSets = {
-  id: string
-  name: string
-}
-
-type DBMetadataStatAffixes = {
-  [key: string]: {
-    id: string
-    affixes: {
-      [key: number]: {
-        affix_id: string
-        property: string
-        base: number
-        step: number
-        step_num: number
-      }
-    }
-  }
-}
-
-export type DBMetadataRelics = {
-  relicMainAffixes: DBMetadataStatAffixes
-  relicSubAffixes: DBMetadataStatAffixes
-  relicSets: Record<string, DBMetadataSets>
-}
-
-export type DBMetadata = {
-  characters: Record<string, DBMetadataCharacter>
-  lightCones: Record<string, DBMetadataLightCone>
-  relics: DBMetadataRelics
-}
 
 // TODO: define specific overrides
 // export type ScoringMetadataOverride = {
