+++ src/lib/tabs/tabRelics/RelicFilterBar.tsx
@@ -8,7 +8,7 @@ import { SettingOptions } from 'lib/overlays/drawers/SettingsDrawer'
 import { RelicScorer } from 'lib/relics/relicScorerPotential'
 import { Assets } from 'lib/rendering/assets'
 import { Renderer } from 'lib/rendering/renderer'
-import DB, { DBMetadataCharacter } from 'lib/state/db'
+import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect'
 import { HeaderText } from 'lib/ui/HeaderText'
@@ -18,6 +18,7 @@ import { Utils } from 'lib/utils/utils'
 import React, { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ReactElement } from 'types/components'
+import { DBMetadataCharacter } from 'types/metadata'
 import { Relic } from 'types/relic'
 
 // FIXME MED
